Rename badly named InnerPage => LoginPanel and move home template to own directory.
