+++ app/controllers/HomeController.scala
@@ -7,7 +7,7 @@ import org.webjars.play.WebJarsUtil
 import play.api.Logger
 import play.api.mvc._
 import play.api.db.Database
-import views.HomeInnerPage
+import views.home.LoginPanel
 
 /**
   * This controller creates an `Action` to handle HTTP requests to the
@@ -29,7 +29,7 @@ class HomeController @Inject() (loginAction: LoginAction, db: Database)(
         s"${routes.ApplicationController.myApplications()}?${request.rawQueryString}"
       )
     else
-      Ok(views.html.home(HomeInnerPage.ConnectionForm))
+      Ok(views.html.home.page(LoginPanel.ConnectionForm))
   }
 
   def status: Action[AnyContent] = Action {
